+++ src/config/index.json
@@ -16,6 +16,8 @@
   "common.confirm": { "th": "ยืนยัน", "en": "Confirm" },
   "common.cancel": { "th": "ยกเลิก", "en": "Cancel" },
   "common.retry": { "th": "ลองอีกครั้ง", "en": "Retry" },
+  "common.viewDetails": { "th": "ดูรายละเอียด", "en": "View details" },
+  "common.hideDetails": { "th": "ซ่อนรายละเอียด", "en": "Hide details" },
   "common.close": { "th": "ปิด", "en": "Close" },
   "common.updating": { "th": "กำลังอัปเดต…", "en": "Updating…" },
   "common.processing": { "th": "กำลังดำเนินการ…", "en": "Processing…" },
@@ -191,6 +193,12 @@
   "account.orders.empty": { "th": "ยังไม่มีคำสั่งซื้อ", "en": "No orders yet." },
   "account.orders.error": { "th": "โหลดคำสั่งซื้อไม่สำเร็จ", "en": "Could not load orders." },
   "account.transactions.empty": { "th": "ยังไม่มีประวัติธุรกรรม", "en": "No transactions yet." },
+  "account.retryUploadSlip": { "th": "ชำระตอนนี้ / อัปโหลดสลิป", "en": "Pay now / Upload slip" },
+  "account.orderItems": { "th": "รายการอาหาร", "en": "Order items" },
+  "account.delivery": { "th": "การจัดส่ง", "en": "Delivery" },
+  "account.distance": { "th": "ระยะทาง", "en": "Distance" },
+  "account.viewOnMap": { "th": "เปิดแผนที่", "en": "View on map" },
+  "account.transaction.expiresAt": { "th": "หมดอายุ", "en": "Expires at" },
   "account.topup.title": { "th": "เติมเงินเข้ากระเป๋า", "en": "Top up balance" },
   "account.topup.description": { "th": "เลือกวิธีเติมเงินเพื่อชำระค่าอาหารได้รวดเร็ว", "en": "Pick a top-up method to pay faster." },
   "account.topup.qr": { "th": "สแกนจ่ายด้วย QR PromptPay", "en": "Scan via PromptPay QR" },
@@ -219,6 +227,8 @@
   "checkout.clearDraft": { "th": "ล้างร่างรายการ", "en": "Clear draft" },
   "checkout.placeOrder": { "th": "สั่งซื้อ", "en": "Place order" },
   "checkout.placeOrderComingSoon": { "th": "สั่งซื้อ (เร็ว ๆ นี้)", "en": "Place order (coming soon)" },
+  "checkout.price": { "th": "ราคา", "en": "Price" },
+  "checkout.addons": { "th": "ตัวเลือกเพิ่มเติม", "en": "Add-ons" },
 
   "line.pageTitle": { "th": "กำลังประมวลผลการเข้าสู่ระบบ LINE…", "en": "Processing LINE sign-in…" },
   "line.statusBoot": { "th": "กำลังเริ่มต้น…", "en": "Booting…" },
