+++ src/constants/i18nKeys.ts
@@ -19,6 +19,8 @@ export const I18N_KEYS = {
     COMMON_CONFIRM: "common.confirm",
     COMMON_CANCEL: "common.cancel",
     COMMON_RETRY: "common.retry",
+    COMMON_VIEW_DETAILS: "common.viewDetails",
+    COMMON_HIDE_DETAILS: "common.hideDetails",
     COMMON_CLOSE: "common.close",
     COMMON_UPDATING: "common.updating",
     COMMON_PROCESSING: "common.processing",
@@ -199,6 +201,12 @@ export const I18N_KEYS = {
     ACCOUNT_ORDERS_EMPTY: "account.orders.empty",
     ACCOUNT_ORDERS_ERROR: "account.orders.error",
     ACCOUNT_TRANSACTIONS_EMPTY: "account.transactions.empty",
+    ACCOUNT_RETRY_UPLOAD_SLIP: "account.retryUploadSlip",
+    ACCOUNT_ORDER_ITEMS: "account.orderItems",
+    ACCOUNT_DELIVERY: "account.delivery",
+    ACCOUNT_DISTANCE: "account.distance",
+    ACCOUNT_VIEW_ON_MAP: "account.viewOnMap",
+    ACCOUNT_TRANSACTION_EXPIRES_AT: "account.transaction.expiresAt",
     ACCOUNT_TOPUP_TITLE: "account.topup.title",
     ACCOUNT_TOPUP_DESCRIPTION: "account.topup.description",
     ACCOUNT_TOPUP_QR: "account.topup.qr",
@@ -229,6 +237,8 @@ export const I18N_KEYS = {
     CHECKOUT_CLEAR_DRAFT: "checkout.clearDraft",
     CHECKOUT_PLACE_ORDER: "checkout.placeOrder",
     CHECKOUT_PLACE_ORDER_COMING_SOON: "checkout.placeOrderComingSoon",
+    CHECKOUT_PRICE: "checkout.price",
+    CHECKOUT_ADDONS: "checkout.addons",
 
     // LINE callback
     LINE_PAGE_TITLE: "line.pageTitle",
